pyproject.toml
[tool.poetry]
name = "notification_sprint"
version = "1"
description = "DmitryK"
authors = ["DmitryK <kachkindmitry@gmail.com>"]
license = "\u001b[A\u001b[B"

[tool.poetry.dependencies]
python = "^3.8"
backoff = "^2.0.1"
pydantic = "^1.9.0"
mypy = "^0.950"
flake8 = "^4.0.1"
orjson = "^3.6.8"
Jinja2 = "^3.1.2"
bitly-api-py3 = "^0.3"
requests = "^2.27.1"
asyncpg = "^0.25.0"
websockets = "^10.3"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
pika = "^1.2.1"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.tox
  | venv
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
multi_line_output = 3


[tool.mypy]
warn_redundant_casts = true
warn_unused_ignores = false
warn_no_return = true
warn_unreachable = true
pretty = true

[[tool.mypy.overrides]]
module = "*.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true


[tool.pylint.messages_control]
max-line-length = 88
disable = [
    "missing-docstring",
    "too-few-public-methods",
    "too-many-arguments",
    "global-statement",
]

ignored-modules = ["pydantic", "orjson"]
good-names = []

[tool.pylint.similarities]
ignore-imports = true
ignore-docstrings = true
ignore-comments = true


[tool.pytest.ini_options]
asyncio_mode= "strict"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(levelname)-5.5s [%(name)s] %(message)s"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
